requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
foreman==0.9.7
gunicorn==23.0.0
honcho==1.1.0
idna==3.9
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-futures==1.0.1
sqlparse==0.5.1
urllib3==2.2.3
whitenoise==6.7.0
